+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/azure/CreateAzureContextFlight.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.service.workspace.flight.create.azure;
+package bio.terra.workspace.service.workspace.flight.azure;
 
 import bio.terra.stairway.Flight;
 import bio.terra.stairway.FlightMap;
