+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/azure/CreateDbAzureCloudContextStartStep.java
@@ -1,41 +1,37 @@
-package bio.terra.workspace.service.workspace.flight;
+package bio.terra.workspace.service.workspace.flight.azure;
 
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
 import bio.terra.workspace.db.WorkspaceDao;
-import bio.terra.workspace.service.workspace.GcpCloudContextService;
-import bio.terra.workspace.service.workspace.exceptions.DuplicateCloudContextException;
+import bio.terra.workspace.service.workspace.AzureCloudContextService;
 import java.util.UUID;
 
 /**
  * Stores the previously generated Google Project Id in the {@link WorkspaceDao} as the Google cloud
  * context for the workspace.
  */
-public class CreateDbGcpCloudContextStep implements Step {
+public class CreateDbAzureCloudContextStartStep implements Step {
   private final UUID workspaceUuid;
-  private final GcpCloudContextService gcpCloudContextService;
+  private final AzureCloudContextService azureCloudContextService;
 
-  public CreateDbGcpCloudContextStep(
-      UUID workspaceUuid, GcpCloudContextService gcpCloudContextService) {
+  public CreateDbAzureCloudContextStartStep(
+      UUID workspaceUuid, AzureCloudContextService azureCloudContextService) {
     this.workspaceUuid = workspaceUuid;
-    this.gcpCloudContextService = gcpCloudContextService;
+    this.azureCloudContextService = azureCloudContextService;
   }
 
   @Override
   public StepResult doStep(FlightContext flightContext) throws InterruptedException {
-    try {
-      gcpCloudContextService.createGcpCloudContextStart(workspaceUuid, flightContext.getFlightId());
-    } catch (DuplicateCloudContextException e) {
-      // On a retry or restart, we may have already started the cloud context create
-    }
+    azureCloudContextService.createAzureCloudContextStart(
+        workspaceUuid, flightContext.getFlightId());
     return StepResult.getStepResultSuccess();
   }
 
   @Override
   public StepResult undoStep(FlightContext flightContext) throws InterruptedException {
     // Delete the cloud context, but only if it is the one we created
-    gcpCloudContextService.deleteGcpCloudContextWithCheck(
+    azureCloudContextService.deleteAzureCloudContextWithFlightIdValidation(
         workspaceUuid, flightContext.getFlightId());
     return StepResult.getStepResultSuccess();
   }
