+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/azure/DeleteAzureContextFlight.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.service.workspace.flight;
+package bio.terra.workspace.service.workspace.flight.azure;
 
 import bio.terra.stairway.Flight;
 import bio.terra.stairway.FlightMap;
@@ -7,6 +7,7 @@ import bio.terra.workspace.common.utils.FlightBeanBag;
 import bio.terra.workspace.common.utils.RetryRules;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequest;
 import bio.terra.workspace.service.job.JobMapKeys;
+import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys;
 import java.util.UUID;
 
 public class DeleteAzureContextFlight extends Flight {
