+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/gcp/CreateCustomGcpRolesStep.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.service.workspace.flight;
+package bio.terra.workspace.service.workspace.flight.gcp;
 
 import static bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.GCP_PROJECT_ID;
