+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/gcp/DeleteGcpContextFlight.java
@@ -1,10 +1,13 @@
-package bio.terra.workspace.service.workspace.flight;
+package bio.terra.workspace.service.workspace.flight.gcp;
 
 import bio.terra.stairway.Flight;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.RetryRule;
 import bio.terra.workspace.common.utils.FlightBeanBag;
 import bio.terra.workspace.common.utils.RetryRules;
+import bio.terra.workspace.service.workspace.flight.DeleteControlledDbResourcesStep;
+import bio.terra.workspace.service.workspace.flight.DeleteControlledSamResourcesStep;
+import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys;
 import bio.terra.workspace.service.workspace.model.CloudPlatform;
 import java.util.UUID;
