+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/gcp/RemoveUserFromWorkspaceFlight.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.service.workspace.flight;
+package bio.terra.workspace.service.workspace.flight.gcp;
 
 import bio.terra.stairway.Flight;
 import bio.terra.stairway.FlightMap;
@@ -8,6 +8,7 @@ import bio.terra.workspace.common.utils.RetryRules;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequest;
 import bio.terra.workspace.service.iam.model.WsmIamRole;
 import bio.terra.workspace.service.job.JobMapKeys;
+import bio.terra.workspace.service.workspace.flight.*;
 import java.util.UUID;
 
 public class RemoveUserFromWorkspaceFlight extends Flight {
