+++ service/src/test/java/bio/terra/workspace/connected/AzureConnectedTestUtils.java
@@ -18,7 +18,7 @@ import bio.terra.workspace.service.resource.controlled.flight.delete.DeleteContr
 import bio.terra.workspace.service.resource.controlled.model.ControlledResource;
 import bio.terra.workspace.service.resource.model.WsmResourceType;
 import bio.terra.workspace.service.workspace.AzureCloudContextService;
-import bio.terra.workspace.service.workspace.flight.create.azure.CreateAzureContextFlight;
+import bio.terra.workspace.service.workspace.flight.azure.CreateAzureContextFlight;
 import java.time.Duration;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
