+++ service/src/test/java/bio/terra/workspace/service/workspace/flight/DeleteControlledAzureResourcesStepTest.java
@@ -23,6 +23,7 @@ import bio.terra.workspace.service.resource.controlled.model.ControlledResource;
 import bio.terra.workspace.service.resource.controlled.model.ControlledResourceFields;
 import bio.terra.workspace.service.resource.controlled.model.ManagedByType;
 import bio.terra.workspace.service.resource.model.CloningInstructions;
+import bio.terra.workspace.service.workspace.flight.azure.DeleteControlledAzureResourcesStep;
 import bio.terra.workspace.service.workspace.model.CloudPlatform;
 import java.util.List;
 import java.util.Optional;
